Resolvee conflict in PR 2152

+++ nczarr_test/run_interop.sh
@@ -69,19 +69,22 @@ case "$zext" in
 	testcasezip ref_quotes zarr metaonly
 	;;
     s3)
-	# Read a test case created by netcdf-java zarr.
-	# Move into position
-        rm -f ${execdir}/ref_zarr_test_data.cdl
-        if gunzip -c < ${srcdir}/ref_zarr_test_data.cdl.gz > ${execdir}/ref_zarr_test_data.cdl ; then
-            testcases3 zarr_test_data.zarr ref_zarr_test_data xarray
+	# Test file does not exist on stratus
+	if test "x$NCZARR_S3_TEST_HOST" = "xs3.us-east-1.amazonaws.com" ; then
+  	    # Read a test case created by netcdf-java zarr.
+	    # Move into position
+            rm -f ${execdir}/ref_zarr_test_data.cdl
+            if gunzip -c < ${srcdir}/ref_zarr_test_data.cdl.gz > ${execdir}/ref_zarr_test_data.cdl ; then
+                testcases3 zarr_test_data.zarr ref_zarr_test_data xarray
+            fi
         fi
 	;;
     *) echo "unimplemented kind: $1" ; exit 1;;
 esac
 }
 
-#testallcases file
-#if test "x$FEATURE_NCZARR_ZIP" = xyes ; then testallcases zip; fi
+testallcases file
+if test "x$FEATURE_NCZARR_ZIP" = xyes ; then testallcases zip; fi
 if test "x$FEATURE_S3TESTS" = xyes ; then testallcases s3; fi
 exit
 # Cleanup
